AEIM-1258 - Add homegrown Puppet function mocking
There are weird problems when trying to mock a function that is defined
in a required module. This approach works around the rspec-puppet-utils
oddities and stubs the DependencyLoader to resolve named Puppet
functions to plain Ruby blocks or RSpec mocks/doubles.

The one requirement imposed on the doubles is that they must mock `call`
to mirror the block/lambda invocation.