AEIM-1293: rewrite throttling & whitelisting
This addresses a couple major issues with the behavior of whitelisted
paths in the previous design that are counter to how throttling
works in the NetScaler implementation:

- If you requested only whitelisted paths, they still incremented the
requst rate. So, if you limited most requests to for example 10 requests
per second, and you did 20 requests to whitelisted paths in one second,
the next request to a non-whitelisted path would get throttled.

- With the temporary ban list in the front end but the whitelists in the
back end, once you had gone over the request rate ALL requests to the
front-end would be denied, even those for whitelisted paths.

Now:

- Throttling only checks the average request rate in the backend -
there's no longer any blocking in the front end (we could bring that
back if desired later while still maintaining the whitelist)

- The whitelist is checked in the front end; there are separate
back-ends for exempted requests. Whitelisted requests don't count
against the limit and will be satisfied even if you've gone over the
request rate

- There is a combined request rate check for http + https requests -
there are separate tracking tables, but the ACL adds the rates and
compares to the limit.