AEIM-1410: Allow adding an "only if" condition for throttling
To minimize disruption to the existing setup with the default backend
throttled and a separate backend exempt from throttling, this is
expressed as a negated condition for the exemption back end - that is,
use the backend exempt from throttling only if the incoming request does
NOT match a particular pattern.

This also changes the optional string parameters to Optional[String]
with a default of undef rather than '' - this seems to be the more
idiomatic way to handle it in puppet given that we still needed
conditionality to handle ''.