AEIM-1502: install i40e-dkms if needed
The custom fact is a little bit janky and depends on pci-utils being
installed; presumably if the fact fails on any given node it just
doesn't have a value there.

If we have additional needs around lspci output we should consider
exposing all the output from lspci -m as a structured custom fact.