AEIM-1429; AEIM-1625: shibboleth configuration
for now, just puts static config files in place
still to do: set permissions so shibd can read SSL keys

The time-out for startup is set by dropping in a configuration file that
is picked up by the systemd service. The value is hardcoded at this
time; should it be expected to change, it should be extracted.