AEIM-1430 - Install PEAR modules with puppet/php
This attempts to install all of the PEAR modules through the
'puppet/php' module, which appears to be the most mature and maintained
on the forge. The usage does not install FPM, dev headers, or composer,
but each is simple to add. I chose to remove the Debian packages for the
select few PEAR modules and list them all as extensions on the PHP
resource.