AEIM-1431: HathiTrust imgsrv
Adds systemd service and startup script for imgsrv.  Parameterizes some
things that were hardcoded in the startup script (sdrroot, sdrview,
sdrdataroot), but removes the parameterization on application.

We could have created a defined type that passed through the application
name as a parameter, but at least for now that likely doesn't make sense
given that we are unlikely to deploy additional Plack-based
applications. We can always revisit that later if need be.