Extract deployment of SSL key & cert to separate profile
This will be in common between all our apache configs.  Some will need
multiple certificates, though, so it probably makes sense to change this
to a defined type when we need that. It would also be nice to not have
to duplicate the full path to the ssl cert & key but I'm not exactly
sure how to do that. It's probably OK to wait until we do AEIM-1236 to
fully think this through, but this is at least a step in (I hope) the
right direction.