AEIM-1628 - Look up staff networks for mod_status
The approach used here does not add another parameter to the Apache
profile, but I would not be opposed to doing so. This commit does an
explicit lookup in hiera. The fixture data includes a model for mapping
networks in multiple places without repeating the actual data, using
alias interpolation.