AEIM-1678: Add ODBC pooling setting
From @skorner's notes:

Pooling is only implemented per process. Shibd is threaded, so pooling is ideal and should be helpful.
Enabling pooling can only be done with Threading=0 or 3, based on whether or not your driver is thread safe.
Otherwise you will see the indefinite hangs, which we have confirmed, and a race condition will occur between
the mutext for pooling and odbc internals. IOW, the odbc lock needs to either be global (3) or within pooling (0).

Also, note that shibd permanently binds a thread to every apache child process. This is why pthread 
mutex issues would be reduced with apache running in worker mode.

Also, note Scott wrote the ODBC plugin to require full serialization and locks on reads.

---

Also:

Best so far: Pooling=Yes and Threading=3. The crash rate is lowest and it is a safe crash mode.