AEIM-1449: Debian repository server role
Includes a "deb signing" profile that deploys a secret key and a script
for signing as well as all dependencies needed to generate the Package
and Releases files and to sign them.

This removes the bare http_fileserver role since the only current use
case is when that file server has a Debian repository.