AEIM-1449: query puppetdb for nodes in other datacenters
The intention is to allow the http file server to permit access to
servers with random IP addresses that aren't in specific datacenter
blocks. This does need an update to the blocks in hiera to list the data
center for each block.

The puppetdb query is pretty ugly, but the results are in an easy to use
format and this allows puppetdb to optimize however it wants rather than
e.g. needing to get a full list of nodes with their data centers, filter
the ones we don't want out, and then query again for the IP addresses. I
didn't see a need to extract the query to a function at this point, but
we could if we need to use it again in a different context.