HT ingest VM: add spec; pass private network template
We do need to have this parameter here so it's overridable via hiera
independently for the role, because the private networking subnet
depends on the environment in a sense (hathitrust vs. general web
hosting, etc).

We might find later on that we want another layer in hiera for this,
probably above the datacenter level, but it means that we might end up
with multiplicative duplicity for things that depend only on the
datacenter and not the environment. But, I'm thinking it might be
better, if for no other reason than we might be able to start pulling
some things out of data/common.yaml. I don't know.

fixup