Make sure apt-update runs before installing packages
- Make sure local apt repo is defined before installing packages from it

- Run an initial apt update and make sure apt-transport-https is
installed before other repos are configured

- Require all apt sources to be applied before installing packages; this
eliminates the need for packages to specify which repository they depend
on, and the possibility that packges could get installed before the repo
is set up. This should be OK as is anyway since packages depend on apt
update, and apt update depends on the sources, but this makes it a
little more explicit.