AEIM-1689: install apt dependencies in preseed
It appeared that on first install the apt caches already existed, but
puppet had removed the /etc/apt/sources.list entries before attempting
to install apt-transport-https and dirmngr. Rather than trying to get
all the ordering and dependencies exactly correct on the puppet side,
this commit simplifies things by installing apt-transport-https and
dirmngr in the preseed and assuming they will be installed ahead of
time.

If this still fails because of issues related to apt-transport-https and
dirmgr (i.e. they're installed but puppet somehow doesn't know that), we
could try removing the dependency of packages on all the apt sources and
just count on the packages depending on the apt update, and that running
after the sources are set up.