AEIM-1760 - Export instances for Apache configuration
This introduces a new resource type to be exported by the named instance
and collected by the appropriate web servers. There are definitely some
rough edges, like the number of values that are shuttled along between
the components, but it's a good start. The template is ported almost
verbatim from ansible-predeploy, with the exception being that the
"extra config" is pruned in favor of "sendfile_path", the one feature of
the inlined config that has been used so far. If we need to model
instances more richly, we can do that, but I hesitate to preserve a
literal expansion from hiera.

The instances in lens will need a number of new keys for this to work.

- Pass correct data structure for create_resources (hash where keys are
the title of the resource, rather than just an array)
- Add required parameters to fixtures in profile/named_instances_spec