split up www_lib vs. www_lib_deepblue
Before, we were relying on the haproxy profile not to produce bindings
for services that weren't configured at that datacenter. This splits up
the www_lib and www_lib + deepblue roles so that www_lib nodes at
Hatcher don't need to also bind to a (nonexistent) deepblue haproxy
service, which seems more correct.