Use exported resources for haproxy node bindings
All haproxy services are virtual resources; we only realize the ones we
actually need, via the haproxy binding resource. Web server nodes export
a haproxy binding with the concat fragments for the node binding and a
realization for the haproxy service node.

The concat fragment resources for node bindings are virtual; that way we
don't always realize the "exemption" backend ones that might not always
be needed. Even though we'll always realize the non-exemption ones,
those are virtual too for consistency.

The haproxy binding resource explicitly passes the hostname, datacenter,
and IP. This is because when we collect the exported resource, facts are
evaluated in the context of the collecting host, not the exporting host.
We might be able to abstract this a little bit by having the nodes
declare an intermediate resource that does the fact collection and
exporting, but I think this is OK for now.

There is still a case where it seems simplest to gather haproxy nodes
via a puppetdb query (for apache). Changing that is out of scope for
now, but we should think about it later to see if we can entirely remove
the puppetdb mocking.