AEIM-1813: turn off enablereuse
with N <= 10workers, imgsrv was timing out after the N+1 request was
issued. This seems to relate to the enablereuse feature -- it looks like
apache was trying to hold open more connections than there were workers
and just timing out. It might in general be fine if we set max to a
lower number than the number of imgsrv workers, but I'd first rather
just do this, especially given eventually we're likely to migrate off of
FastCGI entirely at which point we could try a different connection
re-use strategy. We can also try connection re-use again if it seems
like performance is lackluster with the new VMs.