AEIM-1741 - Add AdoptOpenJDK for tools.lib
- Add a new profile for adoptopenjdk
- Include that profile in the tools_lib role

There are some seemingly extraneous changes here to account for a
problem with rspec-puppet-facts vs. puppetlabs/java, in that
os.architecture is used by the java module and rspec-puppet-facts was
only populating the top level architecture key. As it turns out, the
amd64 value in default_facts was not actually being applied, so it is
removed and we hooked into add_custom_fact in the spec_helper to
normalize the two values.