AEIM-1781: support config for multiple solr cores
Splits out the solr core config to a separate defined type

Adds additional test coverage for all the various parameters & defaults.

Uses ensure_resource for various file things that are used by the solr
core, only necessary if there is at least one solr core, but are really
under control of the named instance. These could alternatively be put
in a conditional block in the named instance, but this is more flexible
in that it allows multiple cores to use the same config directory from
the app, but have a third core that uses a different config directory.