AEIM-1762, AEIM-1859: include solr params in moku init
This changes the flow of information:

- moku node has named_instance resources that:
  - creates concat fragment resources for the moku config on the moku node
  - creates solr_param resources for the moku config on the moku node
  - export 'named_instance::app', 'named_instance::proxy', and
'named_instance::solr_core' resources

This ensures there is a clear responsibility for the flow of
information and that each piece of information has a single resource
responsible for it.

At the named_instance level, some of the parameters used only by the app
or proxy are bundled into app and proxy hashes, which helps reduce the
number of parameters in named_instance and clarify the purpose of some
of the parameters for named instances in hiera.

Also, this changes all json content to use to_json. I think without all
the escaped quotes this makes it both more consistent and more readable.