AEIM-1932: symlink physical core location to solr home
In production, the cores are at e.g. /somewhere/solr/cores, but they
need to be symlinked to the solr home under /var/lib/solr-whatever/home
in order to be picked up at restart.

This adds a separate parameter for the "core home" vs. the "solr home"
and creates that symlink. It also adds some documentation covering how
the solr core defined type works.