Fix failing test for tidying reports
The problem was created in this commit:
c0e590c868d7965aebd122bbd37466f2c0305bf7

I didn't notice the second test that explicitly tested for an age of 1w
instead of 1h, and so we got a failing test. We didn't notice because
changes made on GitHub's UI (as opposed to locally on your machine)
appear not to trigger Travis runs.