AEIM-2055 - Handle apps in user home dir or separate
When a named_instance::app should be deployed to the app user's home
directory, do not create a separate file resource.

Both directories (or just the home directory if the same) should be mode
0750, since the app user should be the only one who needs to write.
However, we are deferring that change because of the dependency on
AEIM-2044. Once that is complete, we can remove the setgid bit and
group-writeable directory business, and rely on moku exec or other
impersonation for writes.