AEIM-2146 - Allow keepalived to bind non-haproxy IPs
This way, we can experiment with alternative HAProxy configurations
before we codify them into puppet formally. Puppet has to manage the
floating IPs because it manages the keepalived config file already.

Without this commit, the only way to get keepalived to bind a floating
IP address was to set up an HAProxy HTTP/S service, which isn't always
what's needed. In the long term, this parameter should be considered a
smell, since it implies some hand-crafted HAProxy configuration that has
yet to be folded into puppet.