PFDR-178: get chipmunk role working
adds default parameters for geoip username & license key so it can be
installed but (effectively) unconfigured. the alternative would be
setting it in the specs in hiera, but it makes no sense to force a hiera
configuration for a component we don't even use for chipmunk. longer
term it would be better to include only the dependencies we actually
need for feed.