AEIM-2129: allow manually adding apt repos
We have had a number of cases where we want to manually add an apt repo
to a server that is at least partially managed by puppet. While we could
add all those extra repos via hiera, I would like to at least for now be
able to disable the automatic purging of apt repos.