AEIM-1999: Cosign configuration (first attempt)
The Cosign directives appear to need to be before the directory config
in order to work, so insert a concat fragment beforehand.

Also, set priority to false so that we can more easily predict the
filename we need to insert the fragment into.

We could consider forking the apache module and adding direct support
for Cosign, but if this works this might be OK until everything can be
migrated to shib.