AEIM-2242 - Fix up tools_lib to succeed on first run
There were two small problems popping up when running on a new AWS VM
provisioned with this role tag:

 1. The mail_recipient had no default value, so we needed a node file
 2. The postgres profile didn't have sufficient dependency information
    on the backup directories

This change introduces a nebula::automation_email to parallel root_email
and a few requires to ensure that postgres is installed before the
backup directories and cron entry are managed.

With a mail_recipient value in common.yml, these hosts can be
provisioned with only an AWS tag.