PFDR-263 - Add PAM/sshd config
This brings in a stock stretch /etc/pam.d/sshd config and introduces an
inline template to include that and then add the umask settings.
Specifically, the umask is set to 0002 on ssh logins, so rsync uploads
can preserve setgid and group write permissions readily with:

rsync -rzvtP --no-group --no-perms --chmod=ugo=rwX

This formula is mentioned in the rsync man page, in the --perms section.