PFDR-274 - Clean up pam/sshd config to avoid repeated corrections
The way the Duo PAM module and the umask settings for Dark Blue were
being managed, they were clobbering each other. This is because the
chipmunk role was overwriting the file, and then the file_line for
pam_duo was modifying it each time.

This commit cleans up how /etc/pam.d/sshd is managed so it is a
concat_file that includes the stock config and then concat_fragments can
add to that the minimal custom file.