AEIM-2593 - Include all but hostname in a VM's domain
This change is not the best:

    -@domain
    +"#{@title}.#{@domain}".split('.').drop(1).join('.')

If it were normal ruby code, I'd want to extract a method as a first
step to clarifying what's going on with this. But this is a template
file, so extracting a method bears an additional cost.

There are tests that represent this behavior, so I'm letting it stand as
is.