Don't manage jruby 1.7
Ruby gems isn't working for jruby 1.7, so we want to purge it entirely.
While we work on that, this ensures there are no failing puppet
resources related to it in the meantime. We could silence the alert, but
then we might miss it if something else starts failing.

These changes should be reverted when 1.7 is gone, as is covered by
AEIM-2776.