AEIM-2815 - Rename kubernetes legacy::kubernetes
This is to make room for its radically different production replacement,
since we want the prototype to continue running until we're ready to
replace it.

This requires a s/kubernetes/legacy::kubernetes/g in the hieradata that
definitely needs to be tested on hatcher-k8s-* before this is safe to
merge.