Use String#include? instead of Regexp#match? in fqdn detection
I'm not sure what version of ruby puppetserver is using, in part because
every version I can find on puppet-master will let me run Regexp#match?
just fine. But in practice, it doesn't look like whatever it uses to
parse erb files can handle it, so I've replaced it with a call to
String#include? instead.