Mount only /www and /www-dev on private app hosts
When just including www_lib::mounts unqualified, the app hosts kept
trying to mount NFS and CIFS filesystems that were www-lib specific and
that they did not have access to. This was causing puppet to raise
errors due to the failed mounts.

This commit explicitly sets those lists of extra mounts to empty hashes
so that hieradata will be ignored.