AEIM-2976 - Check haproxy config on every sync and when reloading
Systemd already checks the config when running reload, but puppet was
running restart instead of reload. This tells it to run reload instead.

Additionally, this adds an exec that checks the config on every sync, so
that, if puppet applies an invalid config, not only will haproxy fail to
reload, but also puppet will repeatedly have this exec as a failing
resource.