AEIM-3213 - Add some new alerts for filling up disks
Three new alerts:

1.  Notice when prometheus isn't collecting metrics
2.  Notice when a disk will fill up within 4 days (based on the past 24 hours)
3.  Notice when a disk will fill up within 4 hours (based on the past hour)

Also tweaked the DeepBlueDataProdStoragePressure alert to give it a
threshold of 3T instead of 2T, since it takes less than 24 hours to go
from 2T free to 0B free.

Also added the hostname to the prometheus config so that the
PrometheusNotRunning alert will produce a helpful error message (without
this, all we'd be able to get is "one of the prometheus servers isn't
working" without any indication as to which).