Add links for clearinghouse conf-available/enabled files
This is a workaround for the time being. The puppetlabs/apache module
expects to manage the conf-enabled directory entirely. This host was set
up with the available/enabled pattern and the links were removed. This
change puts them back, even though we are primarily managing the files
directory now, rather than links -- to be repaired after upgrades
are complete.