AE-8921 - Finish extracting Kerberos from AFS
Here we include the krb5 profile explicitly in all of the places where
the AFS profile was used and implicitly included Kerberos. This is a
decision to "overspecify" in the other profiles and tests on the basis
that we know we want to remove AFS from some of our hosts, and to do so
will be simply deleting the AFS references, rather than applying more
involved changes.