SYS-177 - Move Ruby checks under profile::ruby
We make one assumption in this patch; that the node exporter/prom file
directory exists. We feel it is reasonable because the node exporter is
included in the minimum role. If the ruby profile is used in any role
that does not extend minimum, we should revisit and add a dependency or
exported/collected resources. As it stands, the complexity does not seem
worth the value of the directory potentially going missing.

- Relocate file resources to nebula::profile::ruby
- Relocate check templates to ruby directory
- Inject the rbenv install_dir in the templates