Alert on OOM-kill instead of on < 250M free memory
Our first attempt at a ram alert was to send a page if a host ever has
less than 250M of available memory. In practice, this was noisier than
we'd hoped, since running out of available memory isn't necessarily
indicative of a problem.

We're replacing the rule with two new rules:

1.  HostOomKillDetected will send a page but then quickly resolve itself
    whenever a host OOM-kills something.
2.  HostUnderMemoryPressure will send an info alert whenever there are
    more than 1,000 major page faults (grep pgmajfault /proc/vmstat) per
    minute. Rather than picking a number of bytes that should be free,
    this detects poor performance likely due to low memory.