Use regexes to match YAML output for concat_fragments
It's not clear what conditions aggravate this, but I was seeing
different output on the exact same version of Ruby on my workstation and
in a container. This is presumably the built-in psych implementation
varying in some way that would be hard to track down.

The regexes achieve the same matching, but are not sensitive to the
whitespace difference.