Add comments to explain unintuitive behavior
This file, which appears to have the single responsibility of wrapping
an HTTP port so that it can handle HTTPS responses is a little more
complicated than it could be. This is because it's the only thing in
nebula as of this writing that uses nginx (instead of apache).

Handling http requests to $webroot, for example, doesn't really make
sense in this profile, but it's cordoned off here precisely to prevent
nginx configuration from leaking elsewhere.