Only strip the domain from hosts ending with .umdl.umich.edu
Now that we have other private domains, alerts that strip all but the
hostname can be very confusing. For example, if "worker-123" isn't
syncing with puppet, are we talking about worker-123.private.us-east, or
are we talking about worker-123.private.germany?

We used to strip the full domain because everything we had ended in the
same domain. Going forward, we'll only strip that particular domain, so
that anything with more meaningful domains can be displayed in full.