Unset Shib headers before setting to prevent spoofing
When we always set the X-Shib header (regardless of whether the env var
from mod_shib was present), any client-supplied header would get wiped
out. When conditionally setting the header, a client-supplied header
could be injected if mod_shib does not supply a value. This is not the
behavior we want.