Persist pushgateway metrics across reboots
We thought that the prometheus server would retain metrics that
pushgateways had forgot, but it does not. When metrics disappear from
the pushgateways, they also disappear from the prometheus servers, which
can result in alerts not getting sent.

So the pushgateways will retain metrics going forward.

From now on, when servers are decommissioned, if they have pushed
metrics, someone will have to run this to wipe them:

    curl -X DELETE http://$PROMETHEUS:9091/metrics/hostname/$HOST

Where $PROMETHEUS is the prometheus server and $HOST is the hostname of
the server being decommissioned.