HELIO-4186 - Add default vhost for monitoring
HAProxy checks /monitor/monitor.pl for backend liveness, and that is set
up in apache::base. However, we did not have a default vhost for
requests via IP/fqdn, so the checks were failing. This adds the default
vhost to the fulcrum+apache role. Ideally, it would be in the profile,
but that would collide with the standard inclusion via misc on general
servers.