HELIO-4186 - Add perl dependencies for monitor.pl
There are a few dependencies in monitor.pl, so we add the perl profile.
This is a bummer because the minimal server is getting progressively
bloated... but, the HAProxy http checks depend on /monitor/monitor.pl.
It's slightly preferable to keep that consistent for the time being than
to change the global HAProxy configs to allow a given service to have a
different check or to have something else running at that request path.