Remove legacy www.lib vhost config and add minimal prod fallback support
Hosting of www.lib moved to Netlify in July 2020, but we did
not purge the local apache config for the www.lib vhost.

We want to maintain a www.lib vhost config that is a viable fallback to
host the current www.lib static site if Netlify hosting is failing
for a long period or we need to find alternate hosting in a hurry.

A copy of the production build is already being stored in S3 as a gihub
action and sync'd to /www nightly.