Set UseCanonicalName
This is required for validating the "target" parameter in Shibboleth
sign on.

This also splits out crms-training to a separate vhost; this is required
so that Shibboleth knows the proper return target rather than having to
differentiate between crms-training.babel.hathitrust.org and
babel.hathitrust.org

This also explicitly enables the setenvif module, which is used
elsewhere.