HathiTrust DEV-345: fix redirect to WAYF in dev
We rarely rely on the SP to use the configured WAYF session initiator;
rather, the applications redirect to the WAYF as necessary. However,
there are cases when we do, especially with Dex. This change ensures
that we redirect to the WAYF for the vhost we're accessing rather than
always redirecting to https://babel.hathitrust.org