Stop managing fileserver directories on the puppetserver
This practice has always been (a) weird and (b) of questionable value.
The idea was that we're talking about the directories that puppet
distributes files from, and the question is where those files come from.
If they're unmanaged, when someone has to copy them by hand when
creating a new puppetserver.

With this code, it was possible to create a new puppetserver and let the
old puppetserver hand it all these files. In practice however, it also
means that most of the time, when there's exactly one puppetserver,
every 30 minutes, the puppetserver checks to make sure that every single
file matches... itself. It's an increasingly expensive operation that
cannot possibly accomplish anything of value.

Now that we're starting to add some certificate renewal automation to
our existing certificate distribution flow, we have a conflict. Now
there are two different file resources trying to claim control over
those certificates.

Deleting all the fileserver directory management just for the sake of a
handful of certs could look like throwing the baby out with the
bathwater, but what I hope I've conveyed here is that there never was a
baby in that bathwater.