Allow Fulcrum vhost servername to be set via Hiera
The name of the sites are still fulcrum-http and fulcrum-https, but this
allows the ServerName to be configured. This is on the basis that the
Apache server is dedicated to "this instance" of the Fulcrum app, so
distinguishing it by name is not required. The design for the combined
role is still that each node is both an Apache server (for assets and
Shibboleth) and app server, with no separate tier. Load balancing is
done across self-contained web+app nodes with other services elsewhere.