Reenable profile::users on sysadmin_box
The problems we were having before were because the all_users key of the
nebula::virtual::users class set for default, "first" merge. When we
tried to override the home directories on one machine, it knocked out
the standard user definitions for prometheus, etc., which stopped the
catalogue from applying or updating. Now that we have that set for deep
merge in common.yml, we can override just the home directory values in a
node file and get the effect we wanted -- to make a "special" host, for
better or worse.