Modify press_fcgi_check to report total restarts
We don't currently know how often this script actually does restart the
press service. With this change, the script will increment a counter on
the pushgateway every time it actually runs systemctl restart press, and
if we're ever interested, we'll be able to see how often it comes up, if
ever.