add containerd config file
Default containerd settings, modulo "SystemdCgroup = true". That and
"disabled_plugins = []" are the critical settings, however the many of the
so-called defaults in this file must be set explicitly or containerd does not
work properly. The output of `containerd config default` is known to change
over time, but it's not clear whether those changes are expected to need to be
imported.