Switch from file to concat_fragment for "simple" certs
It is apparently not possible to use a file resource to copy something
from the fileserver that is symlinked there, even with links => follow.

The other parts of these profiles use concat_fragments, and I used file
because I didn't know why. It appears that it is necessary, and this PR
aligns the simple cert process with the others.