Add kubelet to sysadmin boxes
This also creates a generic kubelet profile outside kubernetes, and much
of kubernetes's kubelet resources have been replaced with this profile.
It's not a complete rewrite, but it's a start.

This should definitely be tested in a development environment before
merging.

On the way, this should also fix the naming collision between two
different prometheus node exporter packages by pinning to a specific
version.