Correctly configure HAProxy to check HTTPS on a TCP frontend
If you use the `ssl` option instead of `check-ssl`, then HAProxy will
attempt to add its own TLS encryption on top of the TCP forwarding
instead of merely letting it pass through. Fortunately, you can use
`check-ssl` instead of `ssl`, to tell it only to use SSL on the HTTP
check, not on the main load balancing.