Increase the DeepBlueDataProdFedoraDown check expr
The check happens every 5 minutes, so it should not be considered to be
failing until at LEAST 5 minutes have passed since last success.

I'm leaving `for: 15m` alone because this is only adding a couple
minutes to that potential delay, and it is already the opinion of this
commit author that 15m is, if anything, too short.