HathiTrust DEV-1303: Add USE_CATPROCIO env var
We want to avoid calling 'sudo catprocio' when it isn't going to work;
with a corresponding change in babel, this checks that the USE_CATPROCIO
env var is set and only does sudo catprocio if so.

Sets it both for apache and the imgsrv startup script -- since we are
separately setting env vars for imgsrv startup it certainly doesn't
hurt, even if some of the env vars may come through from apache.