HathiTrust DEV-1318: limit access to metrics endpoint
The metrics endpoint doesn't need to be public, and won't get consistent
data coming from it anyway. This limits it to a provided set of IP
addresses. While we could try to collect IPs for the prometheus role and
the kubernetes gateway, it wasn't immediately clear how to collect the
floating IP that the kubernetes gateway uses, so this defers to a set of
allowed IP addresses listed in lensoftruth.