Stop fighting calico and fix firewall resources
The newer firewall can get into a weird state when port ranges are
defined with hyphens, because in iptables they are defined with colons,
so puppet agent will view this as a corrective change.

This is not necessarily a problem, but when the host's networking is
sufficiently chaotic, this attempt at a corrective change can cause the
puppet server to lose the resource in favor of `nil` and crash.

Calico is sufficiently chaotic, and this commit also aims to stop
purging half of calico's firewall on every puppet agent run.